Adding a "true" gcp likelihood omcputation using the mp_float high-precision numbers. Also added a test using boost.test for the gcp likelihood nad sampling (both seem to fail). Also changed some size_t to int to stop warnings

+++ src/distribution_utils.cpp
@@ -419,7 +419,7 @@ namespace probability_core {
     std::ostringstream oss;
     oss << "{ \"object_class\" : \"gaussian_distribution_t\" , ";
     oss << "  \"mean\" : [ ";
-    for( size_t i = 0; i < gauss.dimension; ++i ) {
+    for( int i = 0; i < gauss.dimension; ++i ) {
       oss << gauss.means[0];
       if( i < gauss.dimension - 1 ) {
 	oss << ",";
